Added tests for duration calculation based on timebase ticks

+++ Firmware/Drivers/Timers/Timer_16_bit/Tests/timer_16_bit_tests.cpp
@@ -563,7 +563,7 @@ TEST(timer_16_bit_driver_tests, test_parameters_computation_prescaler)
     uint32_t cpu_freq = 16'000'000;
     uint32_t target_freq = 1'000;
     uint16_t ocra = 0;
-    uint32_t accumulator = 0;
+    uint16_t accumulator = 0;
 
     timer_16_bit_prescaler_selection_t prescaler = TIMER16BIT_CLK_PRESCALER_1;
     timer_16_bit_compute_matching_parameters(&cpu_freq, &target_freq, &prescaler, &ocra, &accumulator);
