+++ Firmware/Drivers/Timers/Timer_8_bit_async/inc/timer_8_bit_async.h
@@ -388,7 +388,7 @@ void timer_8_bit_async_compute_matching_parameters(const uint32_t * const cpu_fr
                                                    const uint32_t * const target_freq,
                                                    timer_8_bit_async_prescaler_selection_t * const prescaler,
                                                    uint8_t * const ocra,
-                                                   uint32_t * const accumulator);
+                                                   uint16_t * const accumulator);
 
 uint16_t timer_8_bit_async_prescaler_to_value(const timer_8_bit_async_prescaler_selection_t prescaler);
 timer_8_bit_async_prescaler_selection_t timer_8_bit_async_prescaler_from_value(uint16_t const * const input_prescaler);
